+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Development
 
 - WFS and WMS resources are now output as a `dcat:DataService` and matching service distribution. See https://www.dcat-ap.de/def/dcatde/2.0/implRules/#modellierung-eines-datenservices.
+- Ensure that there is always a maximum of one `locn:geometry` per `dct:Location`. It should be the one with datatype `gsp:wktLiteral`.
 
 ## [0.3.7](https://github.com/berlinonline/ckanext-dcatde_berlin/releases/tag/0.3.7)